Add config proxy
It occurred to me that we need to modify the appConfig object so
that it will find css / js that hasn't already been deployed. This
is a work in progress. It seems to be working for the most part, but
there's an odd XML parsing error in the console. Need to investigate...