Remove config module
Originally, the config service was intended to function as a view
properties mapping table in order to facilitate feature flagging,
but its responsibilities have grown and become unwieldy. Now that
we're moving to VE, which does not have an editable view properties
table, it makes sense for components to use the translation service
directly in order to retrieve values stored in the view labels code
table.