Add browser targets for babel
By default @babel/preset-env will transpile everything to ES5 in
order to target the oldest possible browsers, some of which are
explicitly not supported by Primo anymore (IE 11). This configuration
represents a trade-off. It targets slightly modern-ish browsers, and
in exchange, the bundle size is reduced by nearly 25%.

For configuration details, see:
  https://github.com/browserslist/browserslist