Update dev proxy
It turns out that there is an undocumented-but-super-useful feature
in the customization package that enables overriding built-in SVG
icons. If the package contains a file named custom-ui.svg, Primo
will associate the file with a `viewSvg` property on the appConfig
customization object. In order to override a Primo built-in icon,
the SVG id attribute in custom-ui.svg must match the Primo icon.
The PrmIcon component will handle the rest...